requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.20
langchain-core==0.1.52
langchain-groq==0.1.3
sentence-transformers>=2.3.0
faiss-cpu==1.8.0
pydantic>=2.7.4
python-multipart==0.0.6
numpy>=1.26.0
python-dotenv==1.0.0

